scrapper/requirements.txt
selenium
beautifulsoup4
psycopg2-binary